arm64: dts: imx8-apalis: use the hsio clocks for pcie and sata
Use the hsio clocks for pcie and sata as NXP does in their device tree.
We can therefore remove the pcie_ext and sata_ext properties and don't
have to add any special properties.

Mainline for i.MX 8 does not support PCIe yet, this patch cannot be
upstreamed as of now.

Upstream-Status: Pending
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>