arm64: dts: imx8mm-verdin: fix cma memory region
The kernel should search the correct cma memory region. This can be
achieved by deleting the property alloc-ranges. However, we used
delete-node instead of delete-property. Because of this the property was
still existing. With this we fix a mismatch between modules with 1024
and 2048 MB of RAM. The 1024 MB modules showed the following message
during boot and only used 32 MB for CMA:
OF: reserved mem: failed to allocate memory for node 'linux,cma'

We use the kernel configuration to define the CMA size and don't use the
linux,cma node in the device tree.

Upstream-Status: Inappropriate [other]
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>