arm64: dts: imx8mp-verdin: fix cma memory region
The kernel should search the correct cma memory region. This can be
achieved by deleting the property alloc-ranges. However, we used
delete-node instead of delete-property. Because of this the property was
still existing. For the Verdin iMX8MP this didn't have any side effects
for modules with only 1GB RAM because NXP sets alloc-ranges to a value
that works for 1GB and 2GB RAM. It had side effects for the Verdin
iMX8MM with 1GB though this is why we update it here as well.

We use the kernel configuration to define the CMA size and don't use the
linux,cma node in the device tree.

Upstream-Status: Inappropriate [other]
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>