Revert "drm/bridge: sn65dsi83: Register and attach our DSI device at probe"
This reverts commit 6ef7ee48765fa3067858d11ecdf3acbc7c19df80.

In the upstream kernel, the DSI host attach must be called in the bridge
probe function to prevent issues with probe ordering. While this has been
resolved in the upstream kernel, the DSI host driver in the
toradex_6.6-2.0-imx downstream kernel still waits for the bridge to be
ready in its probe function. As a result, when the bridge probes and
attempts to attach to the DSI host, and the DSI host is simultaneously
waiting for the bridge in its probe function, a deadlock occurs due to
probe deferring. This issue affects other bridge drivers because the
upstream changes were made in the same patch series [1]. To clarify, the
problem lies not with the bridge drivers, but with the DSI host driver.

The proper solution is to fix the DSI host downstream driver, but that
would be more complex. The practical solution is to revert the bridge
driver to attach to the DSI host driver during the bridge attach function.
The bridge attach function is called by the DSI host only when the bridge
is ready, ensuring there is no deferred probe.

[1] https://lore.kernel.org/lkml/20210910101218.1632297-1-maxime@cerno.tech/

Upstream-Status: Inappropriate [other]
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>