Revert "drm/imx: imx8mp-ldb: return EPROBE_DEFER in probe"
This reverts commit 445ecaae40ccbe51cbfa0e38650dbddbd081b66e.

This change is no longer required on NXP downstream kernel
v6.6-2.0.x-imx. This change breaks the native LVDS functionality
with the aforementioned kernel version.
Reverting this commit makes the native LVDS work on imx8mp with
kernel v6.6-2.0.x-imx while the original issue of the deferral loop
when native HDMI and the ldb bridge are enabled is not reproducible
anymore. Most likey, below patch series [1] which is part of v6.6
downstream kernel related to fw_devlink may have solved the deferral
loop issue for the mentioned scenario.

[1]: https://lore.kernel.org/all/20240202095636.868578-1-saravanak@google.com/

Upstream-Status: Inappropriate [Other]
Signed-off-by: Parth Pancholi <parth.pancholi@toradex.com>