arm64: dts: imx8mp-verdin: fix gpu memory allocation
There are three methods to allocate memory for the GPU:
- memory-region
- contiguous_mem
- system allocator

The first reserves a fixed memory area for GPU,
the second uses CMA, and the third one act as a fallback
when the others are insufficient albeit with reduced performance.

The current gpu memory-region falls into non-existent physical memory
addresses on boards with less than 2 GB of memory.
Additionally, reserving a memory-region on system with less of 2 GB
reduces the overhall memory available.
So, remove the memory-region usage and rely on contiguous_mem
for GPU memory allocation.

This commit is specific to downstream kernel using galcore driver
instead of etnaviv driver.

Upstream-Status: Inappropriate [Other]
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>