arm64: dts: ti: k3-am62-verdin: Add missing cfg for TI IPC Firmware
The wkup_r5fss0_core0_memory_region is used to store the text/data
sections of the Device Manager (DM) firmware itself and is necessary for
platform boot. Whereas the wkup_r5fss0_core0_dma_memory_region is used
for allocating the Virtio buffers needed for IPC with the DM core which
could be optional. The labels were incorrectly used in the
k3-am62-verdin.dtsi file. Correct the firmware memory region label.

Currently, only mailbox node is enabled with FIFO assignment for a
single M4F remote core. However, there are no users of the enabled
mailboxes. Add the missing carveouts for WKUP R5F and MCU M4F remote
processors, and enable those by associating to the above carveout and
mailboxes. This config aligns with other AM62 boards and can be
refactored out later.

This backport includes a small change:
The mailbox for the Cortex-R5 running the DM firmware has been removed.
The current DM firmware includes an echo RPMsg demo, as noted by TI [1].
This firmware communicates with the Cortex-M4 using hardware FIFO number
5 — a direct channel between the M4 and DM R5 that is not managed by the
Linux kernel.  As a result, when the mailbox is enabled, it
automatically sends a message to the Cortex-M4 even if no firmware is
running on it.

A proper fix would involve either updating the DM firmware or modifying
the omap-mailbox driver to handle only the FIFOs under Linux control.
However, since the DM R5 is not used as a general-purpose remote
processor (the M4F core is) there is no reason to keep this mailbox
enabled at the moment. Keeping it disabled avoids regressions in
suspend/resume behavior. For now, this will remain disabled until a
proper implementation for suspend/resume reaches mainline.

[1] https://e2e.ti.com/support/processors-group/processors/f/processors-forum/1557295/am62p-mailbox-channel-is-not-freed-during-r5-remoteproc-stop-call/5996650

Upstream-Status: Backport [a49f991e740f5f3917b4023705568aeb817ee773]
Signed-off-by: Beleswar Padhi <b-padhi@ti.com>
Tested-by: Hiago De Franco <hiago.franco@toradex.com> # Verdin AM62
Acked-by: Francesco Dolcini <francesco.dolcini@toradex.com>
Link: https://patch.msgid.link/20250908142826.1828676-17-b-padhi@ti.com
Signed-off-by: Nishanth Menon <nm@ti.com>