remoteproc: k3-{m4|r5}: Check mailbox before sending suspend message
During the suspend routine, the mailbox may already have been freed. In
this case, calling mbox_send_message() during suspend returns -EBUSY,
causing the system to fail to enter suspend.

Add a check to ensure that the mailbox (kproc->mbox) is valid before
attempting to send the suspend message. This allows the system to
suspend properly again when the remoteproc node is enabled.

Mainline does not yet have suspend/resume support for the TI K3
remoteproc driver. Therefore, this change is not suitable for upstream.

For linux-ti, the patch was shared with TI using the e2e forum.

Upstream-Status: Submitted [https://e2e.ti.com/support/processors-group/processors/f/processors-forum/1557295/am62p-mailbox-channel-is-not-freed-during-r5-remoteproc-stop-call/6069413]
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>