remoteproc: k3-{m4|r5}: Check rproc state before entering suspend
If the remote core is not running, the suspend routine should return
success immediately instead of sending a message to the remote core and
proceeding with the suspend procedure. Sending suspend messages when the
remote processor is not running makes no sense and can leave a message
pending in the mailbox FIFO, preventing the system from entering
suspend:

omap-mailbox 29010000.mailbox: fifo 1 has unexpected unread messages

Fix this by checking the remote core state before continuing into
suspend.

This is the behavior in the linux-ti downstream 6.12.y branch [1]. Since
mainline currently lacks suspend/resume support for the TI K3 remoteproc
driver, this change is not suitable for upstream.

For linux-ti, the patch was shared with TI using the e2e forum.

[1] https://git.ti.com/cgit/ti-linux-kernel/ti-linux-kernel/tree/drivers/remoteproc/ti_k3_common.c?h=ti-linux-6.12.y#n655

Upstream-Status: Submitted [https://e2e.ti.com/support/processors-group/processors/f/processors-forum/1557295/am62p-mailbox-channel-is-not-freed-during-r5-remoteproc-stop-call/6069413]
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>