Reapply "remoteproc: k3-m4: Move suspend to suspend_late"
This reverts commit 30f40c9d81341e60e8f17141f18a15e2869db04f.

Reverting this commit introduced a regression when the system comes out
of deep sleep:

------------[ cut here ]------------
Unbalanced cpu hotplug enable
WARNING: CPU: 1 PID: 423 at /kernel/cpu.c:563 __cpu_hotplug_enable+0x48/0x58

Instead of reverting this commit, reapply it to fix the regreession so
suspend to RAM can be fixed with a proper solution later.

This reapplyies a commit that is only present on linux-ti downstream
branch, therefore this does not exist upstream and it is not suitable
for mainline Linux kernel.

Upstream-Status: Inappropriate [other]
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>