arm64: dts: ti: k3-am69-aquila: Fix cold reset
On rebooting, a cold reset was not happening as expected. The issue was
that although the COLD_RESET_REQ signal is controlled by software, Linux
has a number of callbacks that can be called on reset, ordered by
priority. However, the callback for the gpio-restart driver, which
controls the COLD_RESET_REQ signal, had less priority than other
handlers in the system and was not being called. Fix this by raising the
priority of the gpio-restart reset handler to 192.

Upstream-Status: Pending
Fixes: da5a1c2da5e1 ("arm64: dts: ti: k3-am69-aquila: add pmic, regulators and control signals")
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>